Implement SomaticHaplotagProcess and Related Classes for Tumor SNP Tagging
- Introduce `SomaticHaplotagProcess`, `SomaticHaplotagBamParser`, and `SomaticHaplotagChrProcessor` to manage tumor-specific SNP tagging.
- Refactor `HaplotagProcess` to support tumor and normal SNP processing with enhanced logging capabilities.
- Implement detailed variant file parsing and chromosome length validation to ensure consistency between tumor and normal datasets.
- Update CIGAR processing logic to accommodate somatic variant handling, improving the overall tagging process.
- Integrate new logging mechanisms in `SomaticTagLog` for comprehensive tracking of tagged reads and variant information.
- Modify `Makefile` to include new source files for proper compilation.
